==> 5_Realtime_data_analysis/S4_Dataframes.ipynb <==
# First contact with df

#### Batch data

Data can be loaded in through a CSV, JSON, XML, or a Parquet file. 

It can also be created using an existing RDD and through any other database, like Hive or Cassandra as well. 

It can also take in data from HDFS or the local file system.

import findspark
import pyspark
import time
import operator
from pyspark import SparkConf
from pyspark import SparkContext
import pandas as pd


conf = SparkConf()
conf.setMaster("local")
conf.setAppName("spark-basic")
sc = SparkContext(conf = conf)

A first example with 'fake' data:

data1 = {'PassengerId': {0: 1, 1: 2, 2: 3, 3: 4, 4: 5, 5: 6 },
         'Name': {0: 'Owen', 1: 'Florence', 2: 'Laina', 3: 'Lily', 4: 'William', 5: 'Jack'},
         'Sex': {0: 'male', 1: 'female', 2: 'female', 3: 'female', 4: 'male', 5: 'male'},
         'Survived': {0: 0, 1: 1, 2: 1, 3: 1, 4: 0, 5: 0}}

data2 = {'Id': {0: 1, 1: 2, 2: 3, 3: 4, 4: 5},
         'Age': {0: 22, 1: 38, 2: 26, 3: 35, 4: 35},
         'Fare': {0: 7.3, 1: 71.3, 2: 7.9, 3: 53.1, 4: 8.0},
         'Pclass': {0: 3, 1: 1, 2: 3, 3: 1, 4: 3}}

df1_pd = pd.DataFrame(data1, columns=data1.keys())
df2_pd = pd.DataFrame(data2, columns=data2.keys())


We will need a spark session to process this type of data. Caution, many examples on the Internet do not specify what "spark." is

from pyspark.sql import SparkSession
spark = SparkSession.builder.getOrCreate()

df1 = spark.createDataFrame(df1_pd)
df2 = spark.createDataFrame(df2_pd)
df1.show()

Summary of the properties of our object

df1.printSchema()

### Operating with dataframes

If you have experience coding with the package dplyr from R, you have more than half of it covered!

**Filter**, **select**, **mutate**, **summarize** and **arrange** are the basic verbs here too, although some of them have different names...

df1.select(["Name","Sex"]).show()

df1.filter(df1.Sex == 'female').filter(df1.Survived == 1).show() 

But we can also do it SQL-style. Note that this is not a python syntax!

df1.filter("Sex='female'").show()

And this is the equivalent of mutate....

hello = df2.withColumn('PricePerYear', df2.Fare/df2.Age)

hello

Groupby is needed when we want to summarize our data... the use is identical than in lists or dictionaries

df2.groupby('Pclass').avg(*["Age","Fare"]).show()

df2.groupby('Pclass').agg({'Age': 'avg', 'Fare': 'avg'}).show()

If you are not familiar with the syntax *****, it refers to "unpack", so we have to provide an unpacked list.

If we want multiple functions at the same time, we need an dictionary

df2.groupby('Pclass').agg({'*': 'count', 'Age': 'avg', 'Fare': 'sum'}).show()

To present the summarized data in a nicer format, we can un toDF() to rename the columns

  (
    df2.groupby('Pclass').agg({'*': 'count', 'Age': 'avg', 'Fare':'sum'})
    .toDF('Pclass', 'counts', 'average_age', 'total_fare')
    .show()
  )

And finally, we can **arrange** dataframes using the method sort

df2.sort('Age', ascending=True).show()

Good reference: https://dzone.com/articles/pyspark-dataframe-tutorial-introduction-to-datafra 

## Basic joins and unions

df1.join(df2, df2.Id == df1.PassengerId).show()

df1.show()

Is it symmetric? Where is Jack? :'(

If you have some time, explore with other types of join, such as "left", "right", "cross", "inner"

#### Your code here

And union doesn't do what we could expect... why would we need it?

df1.union(df2).show()

Some weird join procedure... what is going on here? Are you able to find why would it be useful? (and dangerous?)

df1.join(df2, df1.PassengerId <= df2.Id).show()

And... what about doing joins with SQL?

If we want to do that, we will need temporary Views! Sounds worse than it actually is

df1.createOrReplaceTempView('df1_temp')
df2.createOrReplaceTempView('df2_temp')

query = '''
    select *
    from df1_temp a, df2_temp b
    where a.PassengerId = b.Id'''

spark.sql(query).show()

from pyspark.sql.functions import round, col

resumen = df2.groupby('Pclass').avg("Age").toDF("Clase","Media")
resumen.select("*",round(col("Media"),2)).show()

## Importing and exporting data

#### Indirectly

prod = pd.read_excel("C:/Users/joangj/OneDrive - NETMIND/BTS/2020/data/FoodMarket.xlsx",sheet_name="Products")
cust = pd.read_excel("C:/Users/joangj/OneDrive - NETMIND/BTS/2020/data/FoodMarket.xlsx",sheet_name="Customers")
purc = pd.read_excel("C:/Users/joangj/OneDrive - NETMIND/BTS/2020/data/FoodMarket.xlsx",sheet_name="Purchases")
sell = pd.read_excel("C:/Users/joangj/OneDrive - NETMIND/BTS/2020/data/FoodMarket.xlsx",sheet_name="Sellers")

prod = spark.createDataFrame(prod)
cust = spark.createDataFrame(cust)
purc = spark.createDataFrame(purc)
sell = spark.createDataFrame(sell)

purc.show()

#### Directly

spark.read.csv("C:/Users/joangj/OneDrive - NETMIND/BTS/2020/data/FoodMarket.csv",sep=";",header = True).show()

How would you solve the problem of the decimal format?

#### Writing the file back again

An easy way, but generates a new folder and extra files...

prod.write.csv("Newdata",header = True) 

Using pandas...

prod.toPandas().to_csv('mycsv.csv')

Let's see how many partitions do we have in our data!

purc.rdd.getNumPartitions() 

new_df = purc.groupBy("Seller").count()

new_df.show()

new_df.rdd.getNumPartitions() 

As you can see the partition number suddenly increases. This is due to the fact that the Spark SQL module contains the following default configuration: spark.sql.shuffle.partitions set to 200.

Depending on your use case, this can be benefitial or harmfull. In this particular case, one can see that the data volume is not enough to fill all the partitions when there are 200 of them, which causes unnecessary map reduce operations, and ultimately causes the creation of very small files in HDFS, which is not desirable.

If we want to change this parameter (and many others) we can do it through the SQLContext...

from pyspark.sql import SQLContext
sqlContext = SQLContext(sc)
sqlContext.setConf("spark.sql.shuffle.partitions", "10")
