==> 5_Realtime_data_analysis/S5_SQL.ipynb <==
! pip install pyspark
! pip install findspark

from pyspark.sql import SparkSession
spark = SparkSession \
    .builder \
    .appName("Semi_Sructured_Data_Analysis") \
    .config("spark.some.config.option", "some-value") \
    .getOrCreate()

### Reading CSV


! wget 'https://covid.ourworldindata.org/data/owid-covid-data.csv'

df = spark.read.csv('data/owid-covid-data.csv', sep= ',', header= True)
df.count()

df.select("continent", "location", "date",
          "total_cases", "total_deaths",
          "total_cases_per_million",
          "population"
         ).show(5)

df.createOrReplaceTempView("COVID")

covid_data.columns

spark.sql("SHOW COLUMNS FROM COVID")

spark.sql("DESCRIBE COVID")

df.createOrReplaceTempView("COVID")
covid_data= spark.sql("SELECT location AS country, date, total_cases,\
            total_cases_per_million, population \
            FROM COVID WHERE location ='India'")
covid_data.show(5)

# Window Function SQL: Think of a use case


### Reading Tab Separated Values (TSV)

! wget 'https://github.com/thepanacealab/covid19_twitter/blob/master/dailies/2021-04-19/2021-04-19_clean-dataset.tsv.gz'
! gzip 'data/2021-04-19_clean-dataset.tsv.gz'

# Some other TSV files to play
#! wget 'https://datasets.imdbws.com/title.ratings.tsv.gz' .
#! wget 'https://datasets.imdbws.com/title.ratings.tsv.gz' .
#! gunzip 'data/title.basics.tsv.gz'
#! gunzip 'data/title.ratings.tsv.gz'

df = spark.read.csv("data/2021-04-19-dataset.tsv", header= True, sep= '\t')
df.count()

df.printSchema()

df.show(4)

### Reading JSON



! wget "https://data.cdc.gov/resource/3apk-4u4f.json"

### Binary Files

spark.read.format("binaryFile").option("pathGlobFilter", "*.png").load("/data")



##### Number of Delays:
1. **Carrier**: The number of delays and cancellations due to circumstances within the airline's control (e.g. maintenance or crew problems, aircraft cleaning, baggage loading, fueling, etc.) in this month.
2. **Late Aircraft**: The number of delays and cancellations caused by a previous flight with the same aircraft arriving late, causing the present flight to depart late in this month.	
3. **National Aviation System**: The number of delays and cancellations attributable to the national aviation system that refer to a broad set of conditions, such as non-extreme weather conditions, airport operations, heavy traffic volume, and air traffic control in this month.
4. **Security**: Number of delays or cancellations caused by evacuation of a terminal or concourse, re-boarding of aircraft because of security breach, inoperative screening equipment and/or long lines in excess of 29 minutes at screening areas in this month.
5. **Weather**: Number of delays or cancellations caused by significant meteorological conditions (actual or forecasted) that, in the judgment of the carrier, delays or prevents the operation of a flight such as tornado, blizzard or hurricane in this month.	

import requests, json

url = 'https://think.cs.vt.edu/corgis/datasets/json/airlines/airlines.json'
resp = requests.get(url=url)
json_string = json.dumps(resp.json())
json_data= json.loads(json_string)

with open('airlines.json', 'w') as json_file:
    json.dump(json_data, json_file)

df = spark.read.json("airlines.json")
df.show(6)

df= spark.read.option("multiline","true").json("airlines.json")
#df.show(6)

# Understanding the schema of the JSON through the dataframe
df.printSchema()


# Subsetting the dataframe to analyse the statistics of # (number) of delays
airport_delay_statistics = df.select("Airport.Code", "Airport.Name", 
                                     "Statistics.# of Delays.Carrier", 
                                     "Statistics.# of Delays.Late Aircraft",
                                     "Statistics.# of Delays.Security", 
                                     "Statistics.# of Delays.Weather")
airport_delay_statistics.printSchema()


# Check which input file has created this dataframe
airport_delay_statistics.inputFiles()


# Creating summary statistics of the dataframe attributes
airport_delay_statistics.select("Carrier", "Late Aircraft",
                                "Security","Weather").describe().show()


# Sorting the airport delay statistics based on the Weather
airport_delay_statistics.select("Code", "Name", "Weather").sort("Weather").show(10)


# Sorting the airport delay statistics based on the Weather
airport_delay_statistics.select("Code", "Name", "Weather").sort("Weather").show(10, truncate= False)

airport_delay_statistics.count()


# Find the list of most affected airports due to delays caused by weather
from pyspark.sql.functions import col, sum, count
airport_delay_statistics.select("Code", "Name", "Weather").groupBy("Code").sum("Weather").show(3)#.sort(col("Weather").desc()).show(5, truncate=False)



# Using groupby to find the list of least affected unique airports
from pyspark.sql.functions import col, sum, count
airport_delay_statistics.select("Code", "Name", "Weather") \
    .groupBy("Code", "Name").agg(sum("Weather") \
    .alias("Weather")).sort(col("Weather")).show(10, truncate=False)



# Using groupby to find the list of most affected unique airports
from pyspark.sql.functions import col, sum, count
airport_delay_statistics.select("Code", "Name", "Weather") \
    .groupBy("Code", "Name").agg(sum("Weather") \
    .alias("Weather")).sort(col("Weather").desc()).show(10, truncate=False)


# Find the list of most affected airports due to delays caused by security
airport_delay_statistics.select("Code", "Name", "Security") \
    .groupBy("Code", "Name").agg(sum("Security") \
    .alias("Security")).sort(col("Security").desc()).show(10, truncate=False)


 airport_delay_statistics.columns

airport_delay_statistics.dtypes

airport_delay_statistics.cov("Carrier", "Weather")

# Creating a summary statistics of the dataframe
airport_delay_statistics.describe().show()

# Creating a summary statistics by without columns having NAs 
airport_delay_statistics.describe().dropna().show()

# Creating a summary statistics by without columns having NAs 
airport_delay_statistics.describe().drop("Code", "Name").show()

### Filtering rows 

* Discuss more about the various logical operations that can take place apart from OR such as AND, >, < etc
* discuss about various programming methods to implement the same operation
* explain how to extract the first row using FIRST
* explain how to extract the first N rows using HEAD
* 

# Find the list of most affected airports due to delays caused by security
# Filtering rows using a WHERE clause
airport_statistics.select("Code", "Name", "Security") \
    .groupBy("Code", "Name").agg(sum("Security") \
    .alias("Security")).sort(col("Security").desc()) \
    .where((col("Code")== "ATL") | (col("Code")== "TPA")) \
    .show(10, truncate=False)

# Find the list of most affected airports due to delays caused by security
# Filtering rows using a FILTER clause
airport_statistics.select("Code", "Name", "Security") \
    .groupBy("Code", "Name").agg(sum("Security") \
    .alias("Security")).sort(col("Security").desc()) \
    .filter((col("Code")== "ATL") | (col("Code")== "TPA")) \
    .show(10, truncate=False)

# Find the list of most affected airports due to delays caused by security
# Filtering rows using a WHERE clause

condition_1= (col("Code")== "ATL")
condition_2= (col("Code")== "TPA")

airport_statistics.select("Code", "Name", "Security") \
    .groupBy("Code", "Name").agg(sum("Security") \
    .alias("Security")).sort(col("Security").desc()) \
    .where(condition_1 | condition_2) \
    .show(10, truncate=False)

airport_delay_statistics.isStreaming

airport_delay_statistics.na

airport_delay_statistics.persist

split_1, split_2= airport_delay_statistics.randomSplit([1.0, 2.0], 24)

split_1.show(3)

split_2.show(3)

# df.replace
# df.repartition, df.repartitionByRange
# df.registerTempTable
# df.rdd
# df.intersect
# df.intersectAll
# df.collect vs df.take vs df.show
# df.hint()
# using explode feature
# createGlobalTempView
# createOrReplaceGlobalTempView
# distinct 
# drop_duplicates
# exceptALL
# fillna
# 
# cube function is most important
