==> 5_Realtime_data_analysis/S6_PySpark_Machine_Learning.ipynb <==
from pyspark.sql import SparkSession
from pyspark.sql.functions import explode
from pyspark.sql.functions import split
from pyspark.sql.functions import *

spark = SparkSession \
    .builder \
    .appName("Session-6") \
    .getOrCreate()

%%capture
! wget 'https://raw.githubusercontent.com/databricks/Spark-The-Definitive-Guide/master/data/retail-data/all/online-retail-dataset.csv'

sales = spark.read.csv('online-retail-dataset.csv', header= True, inferSchema= True)
sales.show(5)

# SQL Transformer
from pyspark.ml.feature import SQLTransformer
basicTransformation = SQLTransformer()\
.setStatement("""
SELECT sum(Quantity), count(*), CustomerID
FROM __THIS__
GROUP BY CustomerID
""")
basicTransformation.transform(sales).show(5)

# RFormula based feature transformation
from pyspark.ml.feature import RFormula

dataset = spark.createDataFrame(
    [(7, "US", 18, 1.0),
     (8, "CA", 12, 0.0),
     (9, "NZ", 15, 0.0)],
    ["id", "country", "hour", "clicked"])

formula = RFormula(
    formula="clicked ~ country + hour",
    featuresCol="features",
    labelCol="label")

output = formula.fit(dataset).transform(dataset)
output.show()

## Feature Assembling


# .. means that we are going to the preceding directory beofre reading the corresponding path
df = spark.read.json("../data/airlines.json")
airport_delays = df.select("Airport.Code", 
                          "Statistics.# of Delays.Security", 
                          "Time.Month",
                          "Time.Year")
airport_delays.show(5)

# Vector Assembler
from pyspark.ml.feature import VectorAssembler
va = VectorAssembler().setInputCols(["Month", "Year"])
va.transform(airport_delays).show(5)

## Working with Continuous features



### Bucketizer



from pyspark.ml.feature import Bucketizer

splits = [-float("inf"), -0.5, 0.0, 0.5, float("inf")]
data = [(-999.9,), (-0.5,), (-0.3,), (0.0,), (0.2,), (999.9,)]
dataFrame = spark.createDataFrame(data, ["features"])
bucketizer = Bucketizer(splits=splits, inputCol="features", outputCol="bucketedFeatures")

# Transform original data into its bucket index.
bucketedData = bucketizer.transform(dataFrame)

print("Bucketizer output with %d buckets" % (len(bucketizer.getSplits())-1))
bucketedData.show()

### Scaling and Normalization


%%capture
! wget 'https://raw.githubusercontent.com/Aditya-Mankar/Diabetes-Prediction/master/diabetes.csv'

df = spark.read.csv("diabetes.csv", header= True)
df= df.select(*(col(c).cast("float").alias(c) for c in df.columns))
va = VectorAssembler().setInputCols(['Pregnancies', 'Glucose', 'BloodPressure'])
df= va.transform(df)
vector_column= df.columns[-1]
df_scale= df.select(col(vector_column).alias("features"))

# Standard Scaler
from pyspark.ml.feature import StandardScaler
sScaler = StandardScaler().setInputCol("features")
sScaler.fit(df_scale).transform(df_scale).show(5, truncate= False)

## Working with Categorical Features

%%capture
! wget 'https://archive.ics.uci.edu/ml/machine-learning-databases/00222/bank.zip'
! unzip 'bank.zip'

bank= spark.read.csv("bank.csv", header= True, inferSchema= True, sep= ';')
bank.select("job", "marital", "education", "y").show(5)

from pyspark.ml.feature import OneHotEncoder, StringIndexer
sringIndexer = StringIndexer().setInputCol("job").setOutputCol("job_x")
bank= sringIndexer.fit(bank).transform(bank)
bank.select("job", "job_x").show(10)

encoder = OneHotEncoder(inputCols=["job_x"],
                        outputCols=["job_ohe"])
model = encoder.fit(bank)
encoded = model.transform(bank)
encoded.select("job", "job_x", "job_ohe").show(10, truncate= False)

## Working with Textual Features

# Sales data
# Tokenization using WhiteSpace
from pyspark.ml.feature import Tokenizer
tkn = Tokenizer().setInputCol("Description").setOutputCol("Tokenized_Description")
tokenized = tkn.transform(sales.select("Description"))
tokenized.show(5, False)

# Tokenization using Hyphen 
from pyspark.ml.feature import RegexTokenizer
rt = RegexTokenizer()\
.setInputCol("Description")\
.setOutputCol("Tokenized_Description")\
.setPattern("-")\
.setToLowercase(True)
rt.transform(sales.select("Description")).show(5, False)

# Stopwords removal
from pyspark.ml.feature import StopWordsRemover
englishStopWords = StopWordsRemover.loadDefaultStopWords("english")
stops = StopWordsRemover()\
.setStopWords(englishStopWords)\
.setInputCol("Tokenized_Description")
stops.transform(tokenized).show(20)

# Creating word combinations- NGrams
from pyspark.ml.feature import NGram
bigram = NGram().setInputCol("Tokenized_Description").setN(2)
bigram.transform(tokenized.select("Tokenized_Description")).show(False)

# TF-IDF
tfIdfIn = tokenized\
.where("array_contains(Tokenized_Description, 'red')")\
.limit(10)
tfIdfIn.show(10, False)

ifrom pyspark.ml.feature import HashingTF, IDF
tf = HashingTF()\
.setInputCol("Tokenized_Description")\
.setOutputCol("TFOut")\
.setNumFeatures(10000)

idf = IDF()\
.setInputCol("TFOut")\
.setOutputCol("IDFOut")\
.setMinDocFreq(2)

idf.fit(tf.transform(tfIdfIn))\
.transform(tf.transform(tfIdfIn))\
.select("Tokenized_Description", "TFOut", "IDFOut")\
.show(10)

## Feature Manipulation

from pyspark.ml.feature import VectorAssembler, StandardScaler

df = spark.read.csv("diabetes.csv", header= True)
df= df.select(*(col(c).cast("float").alias(c) for c in df.columns))

# Vector Assembler to assemble all the features into a single vector
va = VectorAssembler().setInputCols(['Pregnancies', 'Glucose', 'BloodPressure'])
df= va.transform(df)
vector_column= df.columns[-1]
df_scale= df.select(col(vector_column).alias("features"))

# Standard Scaler to scale the data
sScaler = StandardScaler().setInputCol("features")
df_scale= sScaler.fit(df_scale).transform(df_scale)
scaled_column= df.columns[-1]
df_scale= df.select(col(scaled_column).alias("scaled_features"))

from pyspark.ml.feature import PCA
pca = PCA().setInputCol("scaled_features").setK(2)
pca= pca.fit(df_scale).transform(df_scale)
pca_features= pca.columns[-1]
pca.select(col("scaled_features"), col(pca_features).alias("principal_components")).show(5, False)

from pyspark.ml.feature import PolynomialExpansion
pe = PolynomialExpansion().setInputCol("scaled_features").setDegree(2)
pe.transform(df_scale).show(5, truncate= False)

## Feature Selection

# Chi-square based feature selection
from pyspark.ml.feature import ChiSqSelector
from pyspark.ml.linalg import Vectors

df = spark.createDataFrame([
    (7, Vectors.dense([0.0, 0.0, 18.0, 1.0]), 1.0,),
    (8, Vectors.dense([0.0, 1.0, 12.0, 0.0]), 0.0,),
    (9, Vectors.dense([1.0, 0.0, 15.0, 0.1]), 0.0,)], ["id", "features", "clicked"])

selector = ChiSqSelector(numTopFeatures=1, featuresCol="features",
                         outputCol="selectedFeatures", labelCol="clicked")

result = selector.fit(df).transform(df)

print("ChiSqSelector output with top %d features selected" 
      % selector.getNumTopFeatures())
result.show()

## Data Pipeline

from pyspark.ml import Pipeline

tkn = Tokenizer().setInputCol("Description").setOutputCol("Tokenized_Description")

englishStopWords = StopWordsRemover.loadDefaultStopWords("english")

stops = StopWordsRemover().setStopWords(englishStopWords)\
.setInputCol("Tokenized_Description")\
.setOutputCol("Removed_Stopwords")

tf = HashingTF()\
.setInputCol("Removed_Stopwords")\
.setOutputCol("TFOut")\
.setNumFeatures(10000)

stages= [tkn, stops, tf]
pipeline = Pipeline(stages = stages)

pipelineModel = pipeline.fit(sales)
df = pipelineModel.transform(sales)
df = df.select("Tokenized_Description", "Removed_Stopwords", "TFOut").show(5)
