==> 5_Realtime_data_analysis/S7_Lecture-7B (Classification).ipynb <==
! wget 'https://archive.ics.uci.edu/ml/machine-learning-databases/00222/bank.zip'
! unzip 'bank.zip'

from pyspark.sql import SparkSession
from pyspark.sql.types import StructType, StructField, IntegerType, StringType, FloatType

spark = SparkSession.builder.master('local[*]').appName('term-deposit-prediction').getOrCreate()
df = spark.read.csv('bank.csv', sep= ';', header = True, inferSchema = True)

df.select(df.columns[:8]).show(4, truncate= False)

# Fidning the number of datapoints in each of the classes
df= df.withColumnRenamed("y", "deposit")
df.groupby('deposit').count().show()

# Extracting a subset of the data for finding summary statistics
df.select(df.columns[:8]).describe().show()

# create a list which has the names of all the numeric features
numeric_features = [t[0] for t in df.dtypes if t[1] == 'int']
df.select(numeric_features).describe().toPandas().transpose()

# removing the features day and month due to non-relevance
df = df.select('age', 'job', 'marital', 'education', 'default', 
               'balance', 'housing', 'loan', 'contact', 'duration', 
               'campaign', 'pdays', 'previous', 'poutcome', 'deposit')

df.dtypes

from pyspark.ml.feature import OneHotEncoder, StringIndexer, VectorAssembler

stages = []
numericCols = [field for (field, dataType) in df.dtypes
               if ((dataType == "int") & (field != "deposit"))]
categoricalColumns = [field for (field, dataType) in df.dtypes
                   if dataType == "string"]
categoricalColumns.remove("deposit")

for categoricalCol in categoricalColumns:
    stringIndexer = StringIndexer(inputCol = categoricalCol, outputCol = categoricalCol + 'Index')
    encoder = OneHotEncoder(inputCols=[stringIndexer.getOutputCol()], outputCols=[categoricalCol + "classVec"])
    stages += [stringIndexer, encoder]

label_stringIdx = StringIndexer(inputCol = 'deposit', outputCol = 'label')
stages += [label_stringIdx]


assemblerInputs = [c + "classVec" for c in categoricalColumns] + numericCols
assembler = VectorAssembler(inputCols=assemblerInputs, outputCol="features")
stages += [assembler]

train, test = df.randomSplit([0.7, 0.3], seed = 2018)
print("Training Dataset Count: " + str(train.count()))
print("Test Dataset Count: " + str(test.count()))

from pyspark.ml.classification import LogisticRegression
from pyspark.ml import Pipeline

lr = LogisticRegression(labelCol="label", featuresCol="features")
stages+= [lr]
pipeline = Pipeline(stages = stages)
pipelineModel = pipeline.fit(train)
pred = pipelineModel.transform(test)
pred.select("features", "label", "prediction").show(5)

from pyspark.ml.evaluation import BinaryClassificationEvaluator
evaluator = BinaryClassificationEvaluator(
    labelCol="label", metricName="areaUnderROC")
accuracy = evaluator.evaluate(pred)
print("Area Under ROC = %g " % (accuracy))

## Random Forest Model

from pyspark.ml.classification import RandomForestClassifier
from pyspark.ml import Pipeline
rf = RandomForestClassifier(featuresCol = 'features', labelCol = 'label')
stages.remove(stages[-1])
stages+= [rf]
pipeline = Pipeline(stages = stages)
pipelineModel = pipeline.fit(train)
pred = pipelineModel.transform(test)
pred.select("features", "label", "prediction").show(5)

from pyspark.ml.evaluation import BinaryClassificationEvaluator
evaluator = BinaryClassificationEvaluator(
    labelCol="label", metricName="areaUnderROC")
accuracy = evaluator.evaluate(pred)
print("Area Under ROC = %g " % (accuracy))

from pyspark.ml.tuning import ParamGridBuilder
paramGrid = (ParamGridBuilder()
    .addGrid(rf.maxDepth, [2, 4, 6])
    .addGrid(rf.numTrees, [10, 100])
    .build())

from pyspark.ml.tuning import CrossValidator
cv = CrossValidator(estimator=pipeline,
    evaluator=evaluator,
    estimatorParamMaps=paramGrid,
    numFolds=3,
    seed=42)
cvModel = cv.fit(train)

list(zip(cvModel.getEstimatorParamMaps(), cvModel.avgMetrics))
