==> 5_Realtime_data_analysis/S9_Streaming_Data.ipynb <==
from IPython.display import display, clear_output
from pyspark.sql import SparkSession
from pyspark.sql import functions as f
import pandas as pd

pd.options.display.max_columns = None
pd.options.display.max_rows = 30
pd.options.display.max_colwidth = 150

! pwd
DATAPATH = "../../../Desktop/RT-activity-data/"

spark = SparkSession.builder.appName("StructuredStreamingApplication").getOrCreate()
schema = spark.read.json(DATAPATH).limit(100).schema

schema

# NOTE: Schema specification is not necessary for a spark DF, but IS necessary for stream
# maxFilePerTrigger is used to specify how many files to pass in each trigger

streaming = spark.readStream.schema(schema).option("maxFilesPerTrigger", 1)\
.json(DATAPATH)

# Selection and filtering
# Streaming query name has to be provided, and should be unique

simpleTransform = streaming.withColumn("stairs", f.expr("gt like '%stairs%'"))\
.where("stairs")\
.where("gt is not null")\
.select("gt", "model", "arrival_time", "creation_time")\
.writeStream\
.queryName("simple_transform")\
.format("memory")\
.outputMode("append")\
.start()

spark.sql("SELECT * FROM simple_transform").show(5)

# Aggregation
deviceModelStats = streaming.cube("gt", "model").avg()\
.drop("avg(Arrival_time)")\
.drop("avg(Creation_Time)")\
.drop("avg(Index)")\
.writeStream.queryName("device_counts").format("memory")\
.outputMode("complete")\
.start()

spark.sql("SELECT * FROM device_counts").show(5)

# spark.sql can be used to request how the query is performing

from time import sleep
for x in range(0, 5):
    display(spark.sql(f"SELECT * from device_counts").toPandas().dropna().head(3))
    sleep(1)
    
    #clear_output(wait=True)
else:
    print("Live view ended...")

static= spark.read.json(DATAPATH)

static.show(4)

# Joins - using a join of a static dataframe on a streaming one
historicalAgg = static.groupBy("gt", "model").avg()
deviceModelStats = streaming.drop("Arrival_Time", "Creation_Time", "Index")\
.cube("gt", "model").avg()\
.join(historicalAgg, ["gt", "model"])\
.writeStream.queryName("device_counts_join").format("memory")\
.outputMode("complete")\
.start()

spark.sql("SELECT * FROM device_counts_join").show(5)

### Triggers

activityCounts = streaming.groupBy("gt").count()

activityCounts.writeStream.trigger(processingTime='5 seconds')\
.format("console").outputMode("complete").start()

activityCounts.writeStream.trigger(once=True)\
.format("console").outputMode("complete").start()

### Event Time and Stateful Processing

spark.conf.set("spark.sql.shuffle.partitions", 5)
static = spark.read.json("./activity-data")
streaming = spark\
.readStream\
.schema(static.schema)\
.option("maxFilesPerTrigger", 10)\
.json("./activity-data")

streaming.printSchema()

withEventTime = streaming\
.selectExpr("*", "cast(cast(Creation_Time as double)/1000000000 as timestamp) as event_time")

# Tumbling Windows
from pyspark.sql.functions import window, col
withEventTime.groupBy(window(col("event_time"), "10 minutes")).count()\
.writeStream\
.queryName("pyevents_per_window")\
.format("memory")\
.outputMode("complete")\
.start()

spark.sql("SELECT * FROM pyevents_per_window").printSchema()

spark.sql("SELECT * FROM pyevents_per_window").show(5)

# Aggregations across multiple columns
from pyspark.sql.functions import window, col
withEventTime.groupBy(window(col("event_time"), "10 minutes"), "User").count()\
.writeStream\
.queryName("events_per_window")\
.format("memory")\
.outputMode("complete")\
.start()

spark.sql("SELECT * FROM events_per_window").show(5)

# Sliding Windows
from pyspark.sql.functions import window, col
withEventTime.groupBy(window(col("event_time"), "10 minutes", "5 minutes"))\
.count()\
.writeStream\
.queryName("pyevents_per_sliding_window")\
.format("memory")\
.outputMode("complete")\
.start()

spark.sql("SELECT * FROM pyevents_per_sliding_window").show(5)

# Watermarks for handling delays
from pyspark.sql.functions import window, col
withEventTime\
.withWatermark("event_time", "30 minutes")\
.groupBy(window(col("event_time"), "10 minutes", "5 minutes"))\
.count()\
.writeStream\
.queryName("watermarking_events")\
.format("memory")\
.outputMode("complete")\
.start()

spark.sql("SELECT * FROM watermarking_events").show(5)

from pyspark.sql.functions import expr
withEventTime\
.withWatermark("event_time", "5 seconds")\
.dropDuplicates(["User", "event_time"])\
.groupBy("User")\
.count()\
.writeStream\
.queryName("pydeduplicated")\
.format("memory")\
.outputMode("complete")\
.start()

spark.sql("SELECT * FROM pydeduplicated").show(5)

### Checkpointing and Fault Tolerance

# Checkpointing
static = spark.read.json("./activity-data")
streaming = spark\
.readStream\
.schema(static.schema)\
.option("maxFilesPerTrigger", 10)\
.json("./activity-data")\
.groupBy("gt")\
.count()

query = streaming\
.writeStream\
.outputMode("complete")\
.option("checkpointLocation", "./checkpointing/")\
.queryName("test_python_stream")\
.format("memory")\
.start()

query.status

query.recentProgress[1]
